Fix 10 bugs, implement Python SDK streaming API
Bug fixes:
- Wire --comms-listen-tcp CLI flag (was silently discarded)
- Deduplicate comms_tool_defs() (surface.rs now uses canonical from dispatcher.rs)
- Fix max_message_bytes type mismatch (usize -> u32, removes as-cast truncation)
- Add DISMISS exit condition to core run_host_mode() (was only in CommsAgent)
- Fix stale model name in comms_verbose example
- Fix hook-authoring SKILL.md (wrong hook point names, wrong count)
- Fix duplicate Python error class hierarchy (generated re-exports from canonical)
- Fix CapabilityEntry.status casing (lowercase -> capital A to match wire format)
- Add capabilities/get to RPC initialize handshake methods list
- Add per-request feature flags to REST CreateSessionRequest

Streaming API:
- _StdoutDispatcher: background multiplexer for rkat rpc stdout (responses -> futures, notifications -> per-session queues)
- StreamingTurn: async context manager + iterable for real-time agent events
- MeerkatClient.create_session_streaming() / start_turn_streaming()
- Catchall queue for create_session where session_id is unknown until response
- 17 new streaming tests, all using mock StreamReader (no process needed)
- Refactored _request() to use dispatcher instead of blocking readline loop
- Deleted broken EventStream class, cleaned up WireEvent phantom fields

+++ meerkat-comms/src/runtime/comms_config.rs
@@ -15,7 +15,7 @@ pub struct CoreCommsConfig {
     pub identity_dir: PathBuf,
     pub trusted_peers_path: PathBuf,
     pub ack_timeout_secs: u64,
-    pub max_message_bytes: usize,
+    pub max_message_bytes: u32,
 }
 
 impl Default for CoreCommsConfig {
@@ -67,7 +67,7 @@ impl CoreCommsConfig {
             trusted_peers_path: resolve(&self.trusted_peers_path),
             comms_config: CommsConfig {
                 ack_timeout_secs: self.ack_timeout_secs,
-                max_message_bytes: self.max_message_bytes as u32,
+                max_message_bytes: self.max_message_bytes,
             },
         }
     }
